[fix] Refactor Twitter to Apple Component.
